% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_conversions}
\alias{coord_conversions}
\title{Build and validate unit conversion factors}
\usage{
coord_conversions(from_units = NULL, to_units = "m")
}
\arguments{
\item{from_units}{Character or NULL; source units.}

\item{to_units}{Character; target units.}
}
\value{
Named list of conversion factors.
}
\description{
Build and validate unit conversion factors
}
\keyword{internal}
