% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_drop_outliers}
\alias{coord_drop_outliers}
\title{Detect and replace outlier coordinates with NA}
\usage{
coord_drop_outliers(output, k = 3)
}
\arguments{
\item{output}{Data frame with x, y columns.}

\item{k}{Numeric; IQR multiplier for outlier fences.}
}
\value{
Modified data frame with outlier x/y/z set to NA.
}
\description{
Detect and replace outlier coordinates with NA
}
\keyword{internal}
