% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_normalise}
\alias{coord_normalise}
\title{Normalise coordinates to an origin}
\usage{
coord_normalise(output, norm)
}
\arguments{
\item{output}{Data frame with x, y, z columns.}

\item{norm}{Logical or numeric; normalisation anchor.}
}
\value{
Modified data frame.
}
\description{
Normalise coordinates to an origin
}
\keyword{internal}
