% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_project}
\alias{coord_project}
\title{Project or convert coordinate data}
\usage{
coord_project(.data, from, to, crs, from_units, to_units, conversions)
}
\arguments{
\item{.data}{A data frame with spatial columns.}

\item{from}{Character; source coordinate type.}

\item{to}{Character; target coordinate type.}

\item{crs}{Numeric; target CRS for latlong projection.}

\item{from_units}{Character or NULL; source units.}

\item{to_units}{Character; target units.}

\item{conversions}{Named list of conversion factors.}
}
\value{
Data frame with x, y, z columns.
}
\description{
Project or convert coordinate data
}
\keyword{internal}
