% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivate.R
\name{derivate}
\alias{derivate}
\title{Calculate Derivatives of Motion}
\usage{
derivate(.data, use_filtered = TRUE, window = 5)
}
\arguments{
\item{.data}{A motion_trace object.}

\item{use_filtered}{Logical: if TRUE, uses f_x and f_y. Defaults to TRUE.}

\item{window}{Numeric: give the size of the window to calculate derivatives across}
}
\value{
A \code{motion_trace} object with velocity and acceleration columns.
}
\description{
Computes linear and angular velocity and acceleration. Defaults to
using filtered coordinates where possible
}
