% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designate.R
\name{designate}
\alias{designate}
\title{designate Motion Data}
\usage{
designate(.data, v_threshold = 1, min_pts = 300, method = "corbett")
}
\arguments{
\item{.data}{Tibble with 'velocity', 'unix_time', and 'is_interpolated'.}

\item{v_threshold}{Velocity (m/s) to distinguish active vs downtime.}

\item{min_pts}{Minimum segment length (e.g., 300 pts = 30s at 10Hz).}
}
\description{
Finalises segmentation using statistical masking and type-safe combining.
}
