% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{.check_extra_collision}
\alias{.check_extra_collision}
\title{Warn on reserved column name collision in user-supplied extra columns}
\usage{
.check_extra_collision(output_names, api_label = "api")
}
\description{
Warn on reserved column name collision in user-supplied extra columns
}
\keyword{internal}
