% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\docType{data}
\name{.mg_pipeline_reserved}
\alias{.mg_pipeline_reserved}
\title{Pipeline-reserved column names produced by derivate() and designate()}
\format{
An object of class \code{character} of length 8.
}
\usage{
.mg_pipeline_reserved
}
\description{
Pipeline-reserved column names produced by derivate() and designate()
}
\keyword{internal}
