% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrate.R
\name{filt_savgol}
\alias{filt_savgol}
\title{Savitzky-Golay filter}
\usage{
filt_savgol(x, window = 5, poly_order = 3)
}
\arguments{
\item{x}{Numeric vector to filter.}

\item{window}{Integer; odd number for the filter window length.}

\item{poly_order}{Integer; polynomial order for the fit.}
}
\value{
Numeric vector of the same length.
}
\description{
Savitzky-Golay filter
}
\keyword{internal}
