% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrate.R
\name{filt_sma}
\alias{filt_sma}
\title{Simple Moving Average filter}
\usage{
filt_sma(x, window = 5)
}
\arguments{
\item{x}{Numeric vector to filter.}

\item{window}{Integer; number of points in the moving window.}
}
\value{
Numeric vector of the same length (with NAs at edges).
}
\description{
Simple Moving Average filter
}
\keyword{internal}
