% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_csv.R
\name{initiate_manual_csv}
\alias{initiate_manual_csv}
\title{Manually Parse CSV Files}
\usage{
initiate_manual_csv(
  .data_path,
  skip = 0,
  col_unix,
  col_lat = NULL,
  col_lng = NULL,
  col_altitude = NULL,
  col_x = NULL,
  col_y = NULL,
  col_z = NULL,
  coord_system = c("gps", "local"),
  max_empty_lines = 3,
  n_max = NULL,
  comment = "#"
)
}
\description{
Manually Parse CSV Files
}
\keyword{internal}
