% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_csv.R
\name{initiate_template_csv}
\alias{initiate_template_csv}
\title{Read a CSV file using a csv_template specification}
\usage{
initiate_template_csv(.data_path, template)
}
\description{
Read a CSV file using a csv_template specification
}
\keyword{internal}
