% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_api.R
\name{load_api_stream_template}
\alias{load_api_stream_template}
\title{Load an API Stream Template from an XML File}
\usage{
load_api_stream_template(path)
}
\arguments{
\item{path}{Character; path to the XML file.}
}
\value{
An \code{api_stream_template} object.
}
\description{
Reads an XML file written by \code{\link{save_api_stream_template}()} and
returns an \code{api_stream_template} object ready to pass to
\code{\link{initiate}()}.
}
\seealso{
\code{\link{api_stream_template}},
\code{\link{save_api_stream_template}}
}
