% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{print.motion_trace}
\alias{print.motion_trace}
\title{Print Method for motion_trace}
\usage{
\method{print}{motion_trace}(x, ...)
}
\arguments{
\item{x}{An object of class \code{motion_trace}.}

\item{...}{Additional arguments.}
}
\description{
Print Method for motion_trace
}
