% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocate}
\alias{allocate}
\title{Allocate Intensity via Named Profiles}
\usage{
allocate(
  .data,
  profile = list(v_bands = c(0, 2, 4, 6, 8), a_bands = c(0, 1, 2, 3, 4), av_bands = c(0,
    100, 200, 300, 400)),
  band_set_name = "profile_1"
)
}
\arguments{
\item{.data}{Tibble with velocity, acceleration, and angular_velocity.}

\item{profile}{A named list containing 'v_bands', 'a_bands', and 'av_bands'.}

\item{band_set_name}{A string to prefix the band columns (default: 'p1').}
}
\description{
Assigns data points to bands with custom names to allow multi-set audits.
}
