% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{create_metadata}
\alias{create_metadata}
\title{Create Metadata}
\usage{
create_metadata(
  session,
  source,
  trace,
  device_info = NULL,
  coord_system = "gps"
)
}
\description{
Create Metadata
}
\keyword{internal}
