% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{fetch_generic_streams}
\alias{fetch_generic_streams}
\title{Fetch and parse data from a generic REST JSON API}
\usage{
fetch_generic_streams(session, template)
}
\description{
Fetch and parse data from a generic REST JSON API
}
\keyword{internal}
