% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrate.R
\name{filtrate}
\alias{filtrate}
\title{Filter Motion Trace Noise on xyz coordinates}
\usage{
filtrate(.data, method = "butterworth", n = 2, cutoff = 0.1, type = "low")
}
\arguments{
\item{.data}{A motion_trace object.}

\item{method}{Character; the filtering algorithm. Just using butterworth right now}

\item{n}{Filter order}

\item{cutoff}{Filter frequency}

\item{type}{filter type (ie., high or low pass)}
}
\value{
A filtered \code{motion_trace} object.
}
\description{
Applies digital filters to x, y, and z coordinates to remove GPS jitter.
}
