% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{get_physics_streams}
\alias{get_physics_streams}
\title{Fetch Physics Streams from Strava}
\usage{
get_physics_streams(activity_id, access_token, start_time_iso, template = NULL)
}
\arguments{
\item{activity_id}{Character; the Strava activity ID.}

\item{access_token}{Character; valid OAuth2 access token.}

\item{start_time_iso}{Character; ISO 8601 start date.}
}
\description{
Fetch Physics Streams from Strava
}
\keyword{internal}
