% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_api.R
\name{get_valid_tokens}
\alias{get_valid_tokens}
\title{Refresh Strava API Tokens}
\usage{
get_valid_tokens(tokens, verbose = TRUE, token_path = "~/strava_tokens.json")
}
\arguments{
\item{tokens}{A list containing client_id, client_secret, refresh_token, and access_token.}

\item{verbose}{Logical; if TRUE, prints status messages.}
}
\value{
A list of updated tokens.
}
\description{
Refresh Strava API Tokens
}
\keyword{internal}
