% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{initiate}
\alias{initiate}
\title{Initiate a Motion Grammar Session}
\usage{
initiate(source = "strava", session = "17134112147", verbose = TRUE)
}
\arguments{
\item{source}{Character; the data provider ('strava' or 'catapult_replay').}

\item{session}{Character; the Strava ID/URL or local file path.}

\item{verbose}{Logical; if TRUE, prints a summary.}
}
\value{
An object of class \code{motion_trace}.
}
\description{
Initiate a Motion Grammar Session
}
