% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.R
\name{initiate}
\alias{initiate}
\title{Initiate a Motion Grammar Session}
\usage{
initiate(
  source = "auto",
  session = "17134112147",
  coord_system = "gps",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{source}{Character; data source ('auto', 'strava', 'catapult_replay', 'guess_csv', 'manual_csv', 'gpx')}

\item{session}{Character; Strava ID/URL or file path}

\item{coord_system}{Character; 'gps' or 'local'}

\item{verbose}{Logical; print summary}

\item{...}{Additional arguments for manual_csv}
}
\description{
Initiate a Motion Grammar Session
}
