% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_csv.R
\name{initiate_guess_csv}
\alias{initiate_guess_csv}
\title{Auto-Detect and Parse Messy CSV Files}
\usage{
initiate_guess_csv(.data_path, source = "gps", confidence_threshold = 0.8)
}
\description{
Auto-Detect and Parse Messy CSV Files
}
\keyword{internal}
